Added Initial Dump Settings to UVM
This adds the intial settings for collecting guest dumps to the uvm.
It configures the guest side registry keys necessary as well as the
document sent to HCS for initially creating the container.

+++ internal/oci/uvm.go
@@ -202,7 +202,7 @@ func handleAnnotationGuestCrashReporting(ctx context.Context, a map[string]strin
 	if !wopts.CrashDumpType.IsValid() {
 		return fmt.Errorf("Invalid crash dump type specified")
 	}
-	wopts.CrashDumpMaxSize = parseAnnotationsUint64(ctx, a, annotations.GuestDumpMaxSize, 0)
+	wopts.CrashDumpMaxSize = int64(parseAnnotationsUint64(ctx, a, annotations.GuestDumpMaxSize, 0))
 	wopts.CrashDumpPath = parseAnnotationsString(a, annotations.GuestDumpLocation, "")
 
 	return nil
